Add iconstore folder
