+++ app/cdap/components/DirectiveInput/IconStore/InfoIcon.tsx
@@ -16,17 +16,6 @@
 
 import React from 'react';
 
-export const CrossIcon = (
-  <svg width="18" height="18" viewBox="0 0 18 18" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <path
-      fill-rule="evenodd"
-      clip-rule="evenodd"
-      d="M0.352464 16.0727C-0.0380608 16.4632 -0.0380608 17.0964 0.352464 17.4869C0.742988 17.8774 1.37615 17.8774 1.76668 17.4869L8.73667 10.5169L15.7067 17.4869C16.0972 17.8774 16.7303 17.8774 17.1209 17.4869C17.5114 17.0964 17.5114 16.4632 17.1209 16.0727L10.1509 9.10271L17.1207 2.13289C17.5112 1.74236 17.5112 1.1092 17.1207 0.718674C16.7302 0.32815 16.097 0.32815 15.7065 0.718674L8.73667 7.6885L1.76685 0.718674C1.37632 0.32815 0.743156 0.32815 0.352632 0.718674C-0.0378928 1.1092 -0.0378924 1.74236 0.352632 2.13289L7.32245 9.10271L0.352464 16.0727Z"
-      fill="white"
-    />
-  </svg>
-);
-
 export const InfoIcon = (
   <svg width="20" height="20" viewBox="0 0 20 20" fill="none" xmlns="http://www.w3.org/2000/svg">
     <circle cx="10" cy="10" r="9" stroke="#79B7FF" stroke-width="2" />
